auth/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
authlib==1.3.0
    # via -r requirements.in
bcrypt==4.1.2
    # via -r requirements.in
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
cffi==1.16.0
    # via cryptography
click==8.1.7
    # via uvicorn
cryptography==42.0.5
    # via authlib
ecdsa==0.19.0
    # via python-jose
exceptiongroup==1.2.1
    # via anyio
fastapi==0.110.2
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via -r requirements.in
idna==3.7
    # via
    #   anyio
    #   httpx
itsdangerous==2.2.0
    # via -r requirements.in
passlib==1.7.4
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via
    #   -r requirements.in
    #   psycopg2-pool
psycopg2-pool==1.2
    # via -r requirements.in
pyasn1==0.6.0
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.7.1
    # via fastapi
pydantic-core==2.18.2
    # via pydantic
python-dotenv==1.0.1
    # via -r requirements.in
python-jose==3.3.0
    # via -r requirements.in
rsa==4.9
    # via python-jose
six==1.16.0
    # via ecdsa
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.29
    # via -r requirements.in
starlette==0.37.2
    # via fastapi
typing-extensions==4.11.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   uvicorn
uvicorn==0.29.0
    # via -r requirements.in
